change api method and fix bugs

+++ pages/api/order/create.ts
@@ -4,6 +4,7 @@ import { METHOD, STATUS_CODE } from "@/const/app-const";
 import { ResponseProps } from "@/network/services/api-handler";
 import { AuthToken } from "@/middleware/server/auth";
 import { prisma } from "@/lib/prisma";
+import { CartDataProps } from "@/contexts/CartContext";
 
 export default async function handler(
   req: NextApiRequest,
@@ -23,35 +24,60 @@ export default async function handler(
 
   const payload = req.body as Order;
   try {
-    const result = await prisma.order.create({
-      data: payload,
-    });
-
-    if (result) {
-      const listCartItems = JSON.parse(result.items) as Cart[];
-      const cartIds = listCartItems.map((item) => item.id);
-
-      await prisma.cart.deleteMany({
-        where: { id: { in: cartIds } },
+    const items: CartDataProps[] = JSON.parse(payload.items);
+    const stockCheck = await Promise.all(
+      items.map(async (item) => {
+        if (item.classificationId !== null) {
+          const classification = await prisma.classification.findUnique({
+            where: { id: item.classificationId },
+          });
+          if (
+            classification?.quantity !== undefined &&
+            classification?.quantity >= item.quantity
+          ) {
+            return true;
+          }
+          return false;
+        }
+        return false;
+      })
+    );
+    if (stockCheck.every((item) => item === true)) {
+      const result = await prisma.order.create({
+        data: payload,
       });
 
-      await Promise.all(
-        listCartItems.map(async (item) => {
-          if (item.classificationId !== null) {
-            return await prisma.classification.update({
-              where: { id: item.classificationId },
-              data: {
-                quantity: { decrement: item.quantity },
-              },
-            });
-          }
-          return;
-        })
-      );
-      return res.status(STATUS_CODE.CREATED).json({
-        code: STATUS_CODE.CREATED,
-        data: result.id,
-        msg: "Tạo đơn thành công",
+      if (result) {
+        const listCartItems = JSON.parse(result.items) as Cart[];
+        const cartIds = listCartItems.map((item) => item.id);
+        await prisma.cart.deleteMany({
+          where: { id: { in: cartIds } },
+        });
+
+        await Promise.all(
+          listCartItems.map(async (item) => {
+            if (item.classificationId !== null) {
+              return await prisma.classification.update({
+                where: { id: item.classificationId },
+                data: {
+                  quantity: { decrement: item.quantity },
+                },
+              });
+            }
+            return;
+          })
+        );
+        return res.status(STATUS_CODE.CREATED).json({
+          code: STATUS_CODE.CREATED,
+          data: result.id,
+          msg: "Tạo đơn thành công",
+        });
+      }
+    } else {
+      return res.status(STATUS_CODE.CONFLICT).json({
+        code: STATUS_CODE.CONFLICT,
+        data: null,
+        msg: "Một hoặc vài sản phẩm đã hết hàng",
       });
     }
   } catch (error) {
